Don't define which helper on Chef 16+ to avoid conflicts
We have our own which helper in the DSL now, which takes 2 args. This
helper is getting injected over it, which works most of the time (not as
well), but doesn't work when someone passes additional args. This breaks
gem_package and probably some other things in the ecosystem. We need to
continue to support chef-sugar being installed on a modern chef-client
release since that's necessary for client upgrades, but we should not
allow the injection of a legacy helper over the built-in.

Signed-off-by: Tim Smith <tsmith@chef.io>